requirements.txt
openai
pandas
python-dotenv
pyyaml
requests
